move handlers related code to `/handlers`
