+++ packages/snaps-utils/src/handlers/settings-page.ts
@@ -1,0 +1,3 @@
+import { OnHomePageResponseStruct } from './home-page';
+
+export const OnSettingsPageResponseStruct = OnHomePageResponseStruct;
