+++ packages/snaps-utils/src/handlers/signature.ts
@@ -1,0 +1,3 @@
+import { OnTransactionResponseStruct } from './transaction';
+
+export const OnSignatureResponseStruct = OnTransactionResponseStruct;
