+++ packages/snaps-utils/src/index.executionenv.ts
@@ -1,6 +1,7 @@
 // Special entrypoint for execution environments for bundle sizing reasons
 export * from './errors';
-export * from './handlers';
+export type { SnapHandler } from './handlers';
+export { HandlerType, SNAP_EXPORT_NAMES } from './handlers';
 export * from './iframe';
 export * from './logging';
 export * from './types';
